requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
djangorestframework==3.14.0
docker==6.1.3
idna==3.4
packaging==23.1
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
urllib3==2.0.4
websocket-client==1.6.1
